==> notebooks/01_load_and_validate_raw_data.ipynb <==
import pandas as pd
from tqdm import tqdm
import datetime
import requests
from pathlib import Path

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """
    Downloads the raw data file for the given year and month from the specified URL.
    Saves the file locally to the 'data/raw' directory with a progress bar.

    Args:
    - year (int): The year of the data file.
    - month (int): The month of the data file.

    Returns:
    - Path: The path where the file was saved locally.
    
    Raises:
    - Exception: If the file cannot be downloaded (e.g., 404 or network error).
    """
    # Construct the URL for the dataset
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"  # Replace with actual base URL

    try:
        response = requests.get(URL, stream=True)
        response.raise_for_status()  # Raise an exception for HTTP errors (e.g., 404, 500)

        # Construct the local file path using pathlib
        save_path = Path(f"../data/raw/rides_{year}-{month:02d}.parquet")

        # Ensure the directory exists
        save_path.parent.mkdir(parents=True, exist_ok=True)

        # Get the total content length (for the progress bar)
        total_size = int(response.headers.get('Content-Length', 0))

        # Write content to the file with a progress bar
        with open(save_path, "wb") as file:
            # Create the tqdm progress bar and iterate over chunks of the response content
            for chunk in tqdm(response.iter_content(chunk_size=1024),
                              total=total_size // 1024,  # Estimate the total number of chunks
                              desc=str(save_path),       # Display the filename in the progress bar
                              unit='KB'):                # Progress unit (Kilobytes)
                file.write(chunk)

        return save_path

    except requests.exceptions.RequestException as e:
        raise Exception(f"Error downloading {URL}: {str(e)}")


download_one_file_of_raw_data(year=2022, month=1)

rides = pd.read_parquet(r"..\data\raw\rides_2022-01.parquet")
rides.head(20)

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id',
}, inplace=True)

rides

# check if all transactions happened in jan 2022
rides['pickup_datetime'].describe()

rides = rides[rides['pickup_datetime'] >= '2022-01-01']
rides = rides[rides['pickup_datetime'] <= '2022-02-01']
rides['pickup_datetime'].describe()

rides.to_parquet("../data/transformed/validated_rides_2022_01.parquet")

## Transform the raw validate data to time series data
