==> notebooks/03_transform_ts_data_into_features_and_targets.ipynb <==
import pandas as pd

ts_data = pd.read_parquet("../data/transformed/ts_data_2022_01.parquet")
ts_data

ts_data_one_location = ts_data.loc[ts_data['pickup_location_id'] == 43, :].reset_index(drop=True)
ts_data_one_location.head()

def get_cutoff_indices(data: pd.DataFrame,
                       n_features: int,
                       step_size: int) -> list:
    
    stop_position = len(data) - 1

    # Start the first sub-sequence at index position 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []

    while subseq_last_idx <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))

        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size

    return indices

n_features = 24
step_size = 1
indices = get_cutoff_indices(
    ts_data_one_location,
    n_features,
    step_size
)

import numpy as np

n_examples = len(indices)
x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
y = np.ndarray(shape=(n_examples), dtype=np.float32)
pickup_hours = []

for i, idx in enumerate(indices):
    x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

print(f'{x.shape=}')
print(f'{x=}')
print(f'{pickup_hours[:5]=}')

features_one_location = pd.DataFrame(x, 
                                     columns= [f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))])
features_one_location

targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])
targets_one_location

from tqdm import tqdm

def transform_ts_data_into_features_and_target(ts_data: pd.DataFrame,
                                               input_seq_len: int,
                                               step_size: int) -> pd.DataFrame:
    
    """
    Slices and transpose data from time-series format into a (feature, target)
    format that we can use to train Supervised ML models
    """

    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    location_ids = ts_data['pickup_location_id'].unique()
    features_list = []
    targets_list = []

    for location_id in tqdm(location_ids):

        # Keep only ts data for this location_id
        ts_data_one_location = ts_data.loc[
            ts_data['pickup_location_id'] == location_id,
            ['pickup_hour', 'rides']
        ].reset_index(drop=True)

        # Pre-compute cutoff indices to split dataframe rows
        indices = get_cutoff_indices(ts_data_one_location,
                                     input_seq_len,
                                     step_size)
        
        # Slice and transpose data into numpy arrays for features and targets
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, input_seq_len), dtype=np.float32)
        y = np.ndarray(shape=(n_examples), dtype=np.float32)
        pickup_hours = []

        rides_values = ts_data_one_location['rides'].values
        pickup_hour_values = ts_data_one_location['pickup_hour'].values

        for i, idx in enumerate(indices):
            x[i, :] = rides_values[idx[0]:idx[1]]
            y[i] = rides_values[idx[1]]
            pickup_hours.append(pickup_hour_values[idx[1]])

        # numpy -> pandas
        features_one_location = pd.DataFrame(x, 
                                     columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(input_seq_len))])
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        # numpy -> pandas
        targets_one_location = pd.DataFrame(y, columns=['target_rides_next_hour'])

        # append to lists instead of concatenating
        features_list.append(features_one_location)
        targets_list.append(targets_one_location)

    # concatenate all at once
    features = pd.concat(features_list, ignore_index=True)
    targets = pd.concat(targets_list, ignore_index=True)

    return features, targets['target_rides_next_hour']

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=27*7*1, # one week of history
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')
