==> notebooks/04_transform_raw_data_into_features_and _targets.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
sys.path.append('../')

from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

import sys
sys.path.append('../')

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

import sys
sys.path.append('../')

import importlib
import src.data
importlib.reload(src.data)

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=27*28*1, # one week of history
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

import sys
sys.path.append('../')

import importlib
import src.paths
importlib.reload(src.paths)

tabular_data = features.copy()
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
